Fix XDG well known directories
Refactored a lot of code to be more modular. Even tho the code in the current state is BAD, it works now without the need to specify a fixed number of well known directories.

This change could also allow to render more menu items on the side bar more easily.