Validate creation of temporary directory in install.sh
The script install.sh creates a temporary directory in order to download
and extract files. We should validate this succeeds otherwise all
following steps can lead to strange results.

Especially since we run `rm -rf` on the temp directory we should ensure
we actually got a temporary directory and are not removing something
unrelated.