fix: fixing `config.toml` (#952)
Before the write, a copy is performed using `io.Copy` and that results
in the write offset being moved to the EOF and the writing to start from
there, effectively resulting in appending to the file instead of
overwriting and that leads to toml decoding errors such as every key
being set twice etc... This fixes that.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Bug Fixes**
* Improved reliability when saving TOML files to ensure updates fully
overwrite the original content.
* Enhanced validation to confirm that fixed TOML configuration files
load correctly without errors.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Nitin Kumar <nknitinkumar12384@gmail.com>